Move some edit logic out of Datasheet.razor.cs
