+++ src/BlazorDatasheet/Datasheet.razor.cs
@@ -569,30 +569,7 @@ public partial class Datasheet : SheetComponentBase
             return;
 
         this.CancelSelecting();
-
-        var cell = Sheet.Cells.GetCell(row, col);
-
-        if (cell.Format.IsReadOnly == true)
-            return;
-
-        // check if the cell is visible OR if the cell is merged, and part of the cell is visible
-        if (!cell.IsVisible)
-        {
-            var mergedRegion = Sheet.Cells.GetMerge(row, col);
-            if (mergedRegion == null)
-                return;
-
-            // is some part of the merge visible?
-            if (Sheet.Rows.GetNextVisible(mergedRegion.Top - 1) > mergedRegion.Bottom)
-                return;
-            if (Sheet.Columns.GetNextVisible(mergedRegion.Left - 1) > mergedRegion.Right)
-                return;
-        }
-
-
-        var softEdit = mode == EditEntryMode.Key || mode == EditEntryMode.None || cell.Value == null;
-
-        Sheet.Editor.BeginEdit(row, col, softEdit, mode, entryChar);
+        Sheet.Editor.BeginEdit(row, col, false, mode, entryChar);
     }
 
     /// <summary>
